+++ packages/quantic/cypress/page-objects/actions/action-set-results-per-page.ts
@@ -1,0 +1,5 @@
+export const setResultsPerPage = () =>
+  cy.get('c-action-results-per-page button').click();
+
+export const setPageSizeValue = (value: number) =>
+  cy.get('c-action-results-per-page input').type(value.toString());