+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.45.9](https://github.com/coveo/ui-kit/compare/headless-react@0.45.8...headless-react@0.45.9) (2021-12-03)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.45.8](https://github.com/coveo/ui-kit/compare/headless-react@0.45.7...headless-react@0.45.8) (2021-12-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.45.7](https://github.com/coveo/ui-kit/compare/headless-react@0.45.6...headless-react@0.45.7) (2021-12-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.45.6](https://github.com/coveo/ui-kit/compare/headless-react@0.45.5...headless-react@0.45.6) (2021-12-01)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.45.5](https://github.com/coveo/ui-kit/compare/headless-react@0.45.4...headless-react@0.45.5) (2021-12-01)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.45.4](https://github.com/coveo/ui-kit/compare/headless-react@0.45.3...headless-react@0.45.4) (2021-12-01)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.45.3](https://github.com/coveo/ui-kit/compare/headless-react@0.45.2...headless-react@0.45.3) (2021-12-01)
 
 **Note:** Version bump only for package headless-react